requirements.txt
imutils==0.5.4
numpy==1.22.3
opencv_python==4.6.0.66
Pillow==9.3.0
